[update] mobile comment rtc
