+++ src/components/Mobile/Entry/EntryFooter/Comment/CommentInputMobile.tsx
@@ -32,6 +32,7 @@ import { useRecoilValue } from "recoil";
 
 interface ICommentInputProps {
   docId: string;
+  refetchFn: () => void;
 }
 
 export const userIcons = [
@@ -69,7 +70,10 @@ export const userIcons = [
   },
 ];
 
-export default function CommentInputMobile({ docId }: ICommentInputProps) {
+export default function CommentInputMobile({
+  docId,
+  refetchFn,
+}: ICommentInputProps) {
   //state
   const colorTheme = useRecoilValue(colorThemeAtom);
   const [userIcon, setUserIcon] = useState<any>(userIcons[0]);
@@ -126,6 +130,7 @@ export default function CommentInputMobile({ docId }: ICommentInputProps) {
     setNickname("");
     setPassword("");
     setComment("");
+    refetchFn();
   };
 
   return (
