+++ src/components/Mobile/Entry/EntryFooter/CommentsMobile.tsx
@@ -1,31 +1,22 @@
 import { Center } from "@chakra-ui/react";
-import { useEffect, useState } from "react";
 import CommentMobile from "./CommentMobile";
-import { fetchComments } from "@/firebase/firebaseUtil";
 import { IComment } from "@/firebase/firebaseTypes";
 
 interface ICommentsProps {
-  docId: string;
+  comments: IComment[];
+  refetchFn: () => void;
 }
 
-export default function CommentsMobile({ docId }: ICommentsProps) {
-  const [comments, setComments] = useState<IComment[] | undefined>(undefined);
-
-  const getComments = async (docId: string) => {
-    const fetchData = await fetchComments(docId);
-    setComments(fetchData);
-  };
-
-  useEffect(() => {
-    getComments(docId);
-  }, [docId]);
-
+export default function CommentsMobile({
+  comments,
+  refetchFn,
+}: ICommentsProps) {
   return (
     <>
       <Center w="full" h={"auto"} flexDir={"column"} gap={4} px={2}>
         {comments?.map((comment) => (
           <CommentMobile
-            key={comment.id}
+            key={"COMMENTMOBILE" + comment.id}
             commentId={comment.id!}
             nickname={comment.nickname!}
             password={comment.password!}
@@ -33,6 +24,7 @@ export default function CommentsMobile({ docId }: ICommentsProps) {
             comment={comment.comment!}
             createdAt={comment.createdAt!}
             edited={comment.edited!}
+            refetchFn={refetchFn}
           />
         ))}
       </Center>
